+++ src/main/java/com/JustHealth/Health/Repository/InventoryRepository.java
@@ -2,6 +2,7 @@ package com.JustHealth.Health.Repository;
 
 import com.JustHealth.Health.Entity.Batch;
 import com.JustHealth.Health.Entity.Inventory;
+import com.JustHealth.Health.Entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +11,8 @@ import java.util.List;
 @Repository
 public interface InventoryRepository extends JpaRepository<Inventory,Long> {
 
-    Inventory findByProductId(Integer productId);
+    Inventory findByProductId(Long productId);
+//    Product findByInventoryId(Long inventoryId);
